refactor: User 관련 도메인 패키지 분리
